% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wfm.R
\name{as.wfm}
\alias{as.wfm}
\title{Coerce to a Word Frequency Matrix}
\usage{
as.wfm(mat, word.margin = 1)
}
\arguments{
\item{mat}{a matrix of counts}

\item{word.margin}{which margin of mat represents the words}
}
\value{
an object of class wfm
}
\description{
Constructs a wfm object from various other kinds of objects
}
\author{
Will Lowe
}
\seealso{
\code{\link{wfm}}
}

