\name{austin-package}
\alias{austin-package}
\alias{austin}
\docType{package}
\title{
\packageTitle{austin}
}
\description{
\packageDescription{austin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{austin}
\packageIndices{austin}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{austin}

Maintainer: \packageMaintainer{austin}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
