% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{package}
\name{austin}
\alias{austin}
\title{austin: Do things with words}
\description{
Austin helps you see what people, usually politicians, do with words.
Currently that means how positions on a presumed underlying
policy scale are taken by manipulating word occurrence counts.  
The models implemented here try to
try to recover those positions using only this information, plus
some heroic assumptions about language generation, e.g.
unidimensionality, conditional independence of words given ideal point
and Poisson-distributed word counts.
}
\details{
The package currently implements Wordfish (Slapin and Proksch, 2008) and 
Wordscores (Laver, Benoit and Garry, 2003).  See references for details.
}
