% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{binary_search}
\alias{binary_search}
\title{Binary search}
\usage{
binary_search(fun, target_value, lower, upper, epsilon = 1e-06, ...)
}
\arguments{
\item{fun}{a function to test the boundary values}

\item{target_value}{the target value}

\item{lower}{a lower boundary}

\item{upper}{an upper boundary}

\item{epsilon}{the minimum interval size of \link{lower,upper}}

\item{...}{}
}
\value{
an approximation of the target value
}
\description{
Binary search
}
