% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.wordscores.R
\name{coef.classic.wordscores}
\alias{coef.classic.wordscores}
\title{Show Wordscores}
\usage{
\method{coef}{classic.wordscores}(object, ...)
}
\arguments{
\item{object}{a fitted Wordscores model}

\item{...}{extra arguments, currently unused}
}
\value{
The wordscores
}
\description{
Lists wordscores from a fitted Wordscores model.
}
\seealso{
\code{\link{classic.wordscores}}
}
\author{
Will Lowe
}
