% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfish.R
\name{coef.wordfish}
\alias{coef.wordfish}
\title{Extract Word Parameters}
\usage{
\method{coef}{wordfish}(object, form = c("poisson", "multinomial"), ...)
}
\arguments{
\item{object}{an object of class wordfish}

\item{form}{which parameterization of the model to return parameters for}

\item{...}{extra arguments}
}
\value{
A data.frame of word parameters from a wordfish model in one or
other parameterization.
}
\description{
Extract word parameters beta and psi in an appropriate model
parameterization
}
\details{
Slope parameters and intercepts are labelled beta and psi respectively.  In
multinomial form the coefficient names reflect the fact that the
first-listed word is taken as the reference category.  In poisson form, the
coefficients are labeled by the words the correspond to.

Note that in both forms there will be beta and psi parameters, so make sure
they are the ones you want.
}
\author{
Will Lowe
}
\seealso{
\code{\link{wordfish}}
}

