% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{daildata}
\alias{daildata}
\title{The Irish No-Confidence debate}
\description{
Irish No-Confidence Debate
}
\details{
This are word counts from the no-confidence debate in Ireland.

\code{daildata} is a word frequency object.
}
\references{
Benoit and Laver's Irish Political Studies piece. (fixme!)
}
\keyword{datasets}
