% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{demanif}
\alias{demanif}
\title{German Party Manifesto Data}
\source{
Wordfish website (http://www.wordfish.org)
}
\description{
A random sample of words and their frequency in German political party
manifestos from 1990-2005.
}
\details{
demanif is word frequency matrix
}
\references{
J. Slapin and S.-O. Proksch (2008) 'A scaling model for
estimating time-series party positions from texts' American Journal of
Political Science 52(3), 705-722.
}
\keyword{datasets}

