% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{estimate_intercept}
\alias{estimate_intercept}
\title{Estimate intercept for given marginal probability}
\usage{
estimate_intercept(candidate, linear_predictor)
}
\arguments{
\item{candidate}{a candidate value for the target intercept}

\item{linear_predictor}{the linear predictor withour the intercept}
}
\value{
marginal probability using the candidate intercept
}
\description{
Estimate intercept for given marginal probability
}
