% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{estimate_marginal_risk_diff}
\alias{estimate_marginal_risk_diff}
\title{Estimate marginal risk difference}
\usage{
estimate_marginal_risk_diff(candidate, linear_predictor, marg_prob_untreated)
}
\arguments{
\item{candidate}{candidate value for marginal risk difference}

\item{linear_predictor}{linear predictor without intercept}

\item{marg_prob_untreated}{intended marginal probability untreated}
}
\value{
marginal risk difference obtained with candidate value
}
\description{
Estimate marginal risk difference
}
