% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wordfish.R
\name{fitted.wordfish}
\alias{fitted.wordfish}
\title{Get Fitted Values from a Wordfish Model}
\usage{
\method{fitted}{wordfish}(object, ...)
}
\arguments{
\item{object}{a fitted Wordfish model}

\item{...}{Unused}
}
\value{
Expected counts in the word frequency matrix
}
\description{
Extracts the estimated word rates from a fitted Wordfish model
}
\author{
Will Lowe
}

