% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate predictor and outcome data}
\usage{
generate_data(
  sample_size = 10000,
  n_covariates = 10,
  n_normal = n_covariates,
  pair_cor,
  alpha,
  beta = NULL,
  outcome_type = NULL,
  n_iter = 1000,
  prop_treated = NULL,
  risk_diff = 0,
  margin_prev
)
}
\arguments{
\item{sample_size}{integer specifying the sample size}

\item{n_covariates}{integer specifying the number of covariates}

\item{n_normal}{integer specifying the number of normal covariates (must be <= n_convariates)}

\item{pair_cor}{pairwise correlation}

\item{alpha}{covariate vector of length \code{n_covariate}}

\item{beta}{coefficient for treatment-dummy}

\item{outcome_type}{character value \code{"continuous"} or \code{"binary"}}

\item{n_iter}{integer indicating number of iterations}

\item{prop_treated}{numerical value of intended proportion of treated individuals}

\item{risk_diff}{numerical value of intended risk difference}

\item{margin_prev}{numerical value of intenden marginal prevalence}
}
\value{
a list with the following content
\itemize{
\item \code{sim_data} dataframe with simulated data}
\item \code{alpha_0_treat} numerical value indicating intercept of treatment model
\item \code{alpha_0_outcome} numerical value indicating intercept of outcome model
\item \code{beta} numerical value indicating regression coefficient of treatment dummy
}
}
\description{
Generate predictor and outcome data
}
