% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{get_cov_matrix}
\alias{get_cov_matrix}
\title{Assemble correlation matrix}
\usage{
get_cov_matrix(n_normal, pair_cor)
}
\arguments{
\item{n_normal}{integer specifying the number of normal covariates}

\item{pair_cor}{pairwise correlation}
}
\value{
correlation matrix of size \code{n_normal}^2
}
\description{
Assemble correlation matrix
}
