% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{get_linear_predictor}
\alias{get_linear_predictor}
\title{Obtain linear predictor}
\usage{
get_linear_predictor(alpha, covariate_data)
}
\arguments{
\item{alpha}{coefficient vector}

\item{covariate_data}{data frame with predictor data}
}
\value{
a vector
}
\description{
Obtain linear predictor
}
