% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{getdocs}
\alias{getdocs}
\title{Get Documents}
\usage{
getdocs(wfm, which)
}
\arguments{
\item{wfm}{a wfm object}

\item{which}{names or indexes of documents}
}
\value{
A smaller wfm object containing only the desired documents with the
same word margin setting as the original matrix.
}
\description{
Gets particular documents from a wfm by name or index
}
\details{
getdocs is essentially a subset command that picks the correct margin for
you.
}
\author{
Will Lowe
}
\seealso{
\code{\link{as.wfm}}, \code{\link{as.docword}},
\code{\link{as.worddoc}}, \code{\link{docs}}, \code{\link{words}},
\code{\link{is.wfm}}, \code{\link{wordmargin}}
}

