% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{iebudget2009}
\alias{iebudget2009}
\alias{iebudget2009cov}
\title{Irish Budget Debate Data 2009}
\description{
Irish budget debate 2009
}
\details{
This are word counts from the 2009 Budget debate in Ireland.

This is a word frequency nmatrix.  Loading this data also makes available
\code{iebudget2009cov} which contains covariates for the speakers.
}
\keyword{datasets}

