% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{interestgroups}
\alias{interestgroups}
\title{Interest Groups}
\description{
Interest Groups and the European Commission
}
\details{
Word counts from interest groups and a European commission proposal to
reduce CO2 emmissions in 2007.

\code{comm1} and \code{comm2} are the commission's proposal before and after
the proposals of the interest groups.
}
\references{
H. Kluever (2009) 'Measuring influence group influence using
quantitative text analysis' European Union Politics 11:1.
}
\keyword{datasets}

