% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_collapse}
\alias{jl_collapse}
\title{Collapse counts to groups}
\usage{
jl_collapse(x)
}
\arguments{
\item{x}{a grouped jl_df}
}
\value{
an jl_df
}
\description{
If a jl_df is grouped, collapse the 'counts' within each group. Like 
summarize, after group_by, but with no choice of function.
}
