% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_count}
\alias{jl_count}
\title{Tabulate tokens}
\usage{
jl_count(x, drop_tokens_after = FALSE)
}
\arguments{
\item{x}{a jl_df object}

\item{drop_tokens_after}{whether to drop 'tokens' variable after 'counts' computation}
}
\value{
a jl_df with a 'counts' variable
}
\description{
The 'tokens' variable is turned into a 'counts' variable 
containing sparse matrices of 
table counts, one per document. Any 'tokens' variable is maintained
unless drop_tokens_after is TRUE.
}
