% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_count_categories}
\alias{jl_count_categories}
\title{Get category counts from a dictionary}
\usage{
jl_count_categories(x, dictionary, ...)
}
\arguments{
\item{x}{a tibble}

\item{dictionary}{a quanteda content analysis dictionary}

\item{...}{extra arguments to tokenizers::tokenizer_*}
}
\value{
a tibble with 'tokens' and 'counts' variables
}
\description{
Runs a dictionary on the text column of x to create 'tokens' variable 
consisting of dictionary matches to each word, (non-matching words
from the dictionary generate no token), then tabulates them within 
each document in 'counts' variable
}
\details{
This is a shortcut for running  jl_tokenize_categories then jl_count_tokens
and should be used in preference to these two.
}
