% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_count_tokens}
\alias{jl_count_tokens}
\title{Tabulate whatever the tokens are}
\usage{
jl_count_tokens(x)
}
\arguments{
\item{x}{a tibble with a 'tokens' variable}
}
\value{
a tibble with a 'counts' variable
}
\description{
Tabulate whatever the tokens are
}
