% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_count_words}
\alias{jl_count_words}
\title{Get word counts}
\usage{
jl_count_words(x, ...)
}
\arguments{
\item{x}{a tibble}

\item{...}{extra arguments to tokenizers::tokenize_*}
}
\value{
a tibble
}
\description{
Splits 'text' into words as a 'tokens' variable then tabulates them 
within each document in 'counts' variable
}
\details{
This is a shortcut for running  jl_tokenize_words then jl_count_tokens
and should be used in preference to these two.
}
