% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_demote_counts}
\alias{jl_demote_counts}
\title{Undo the effects of jl_promote_counts}
\usage{
jl_demote_counts(x, prefix = NULL)
}
\arguments{
\item{x}{a tibble}

\item{prefix}{whether to look for a prefix when removing}
}
\value{
a tibble
}
\description{
Removes the variables that jl_promote_counts added
}
