% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_dfm}
\alias{jl_dfm}
\title{Extract a document feature matrix}
\usage{
jl_dfm(x, rownames_from = "doc_id", inflate = FALSE)
}
\arguments{
\item{x}{a tibble}

\item{rownames_from}{which variable to take the row names from (default: doc_id)}

\item{inflate}{whether to return a dense base::matrix (TRUE) or a Matrix::sparseMatrix}
}
\value{
a matrix
}
\description{
Throws away all variables and constructs a documnt feature 
matrix with column names from the types of whatever is counted in 'counts'
ans row names from a chosen variable, or doc_id.
}
\details{
Note that requesting an inflated document feature matrix turns a 
Matrix::sparseMatrix into a regular base::matrix and may increase 
memory consumption considerably.
}
