% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_filter}
\alias{jl_filter}
\title{Filter a corpus}
\usage{
jl_filter(x, ..., .preserve = FALSE)
}
\arguments{
\item{x}{a tibble}

\item{...}{a dplyr expression specifying rows}

\item{.preserve}{whether to keep groups (default: FALSE)}
}
\value{

}
\description{
Filter out some documents using dplyr expressions, reindexing 
'counts' as necessary
}
