% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_freqs}
\alias{jl_freqs}
\title{Get type counts for tokens}
\usage{
jl_freqs(x)
}
\arguments{
\item{x}{a jl_df object}
}
\value{
a vector of type counts
}
\description{
This assumes that a 'counts' variable exists. Use jl_count to create it
if necessary.
}
