% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_identify}
\alias{jl_identify}
\title{Add a document identifier}
\usage{
jl_identify(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a tibble
}
\description{
Adds a variable 'doc_id' to uniquely identify each row. This identifier
may later, if jl_split is used, contain information about disaggregation 
level. If the document identifier already exists, overwrite it.
}
