% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_incorporate}
\alias{jl_incorporate}
\title{Create a corpus}
\usage{
jl_incorporate(x, text_var = "text", remove_old_text_vars = TRUE)
}
\arguments{
\item{x}{a data frame or tibble}

\item{text_var}{the name of the variable containing text}

\item{remove_old_text_vars}{whether to remove old text variable after renaming}
}
\value{
a tibble
}
\description{
Make a data frame or tibble fit for text analysis.
}
