% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_inflate}
\alias{jl_inflate}
\title{Divide each texts}
\usage{
jl_inflate(x, tokenizer = tokenizers::tokenize_paragraphs, ...)
}
\arguments{
\item{x}{a tibble}

\item{tokenizer}{a function that splits each text}

\item{...}{extra arguments to give to tokenizer}
}
\value{
a tibble with new doc_id
}
\description{
This function inflates the jl_df by 'exploding' each text using the 
tokenizer function and duplicating the non-textual variables across
the new elements. By default tokenizer choice this function makes 
each paragraph a new document. Replace this function with your prefered 
splitting function. It should return a list of character vectors. To 
add extra arguments to the tokenizer function call, put them in ...
}
