% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_lengths}
\alias{jl_lengths}
\title{Get document lengths in tokens}
\usage{
jl_lengths(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a vector of token counts
}
\description{
For each document, how many tokens it contains. Note that this may 
not correspond to the number of words or dictionary categories
}
