% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_process_topics}
\alias{jl_process_topics}
\title{Create corpus from wide counts}
\usage{
jl_process_topics(x, ..., drop_old_vars = TRUE, drop_unused_vars = TRUE)
}
\arguments{
\item{x}{a tibble}

\item{...}{a tidy select expression}

\item{drop_old_vars}{whther to remove the wide variables (default: TRUE)}

\item{drop_unused_vars}{whether to drop wide variables whose column sums are zero}
}
\value{
a tibble with a 'counts' variable
}
\description{
Given a tibble x of variables, some of which are wide form category counts, and 
the remainder of which are document variables (docvars in quanteda terminology),
decompose the variables specified by the tidyselect arguments into a 'counts'
variable in sparse format and keep the remainder.
}
