% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_promote_counts}
\alias{jl_promote_counts}
\title{Make counts full variables in wide form}
\usage{
jl_promote_counts(x, prefix = NULL)
}
\arguments{
\item{x}{a jl_df with 'counts'}

\item{prefix}{what to prefix each counted element's value when it turns into a variable name}
}
\value{
a jl_df
}
\description{
Make counts full variables in wide form
}
