% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_split}
\alias{jl_split}
\title{Split a corpus into subunits}
\usage{
jl_split(x, what = c("paragraphs", "sentences", "regex"), ...)
}
\arguments{
\item{x}{a tibble}

\item{what}{what unit to disaggregate a document to (default: paragraphs)}

\item{...}{extra arguments to give to tokenizers::tokenize_*}
}
\value{
a tibble with new doc_id
}
\description{
If a document 23 of x contains 3 sentences in its 'text' then 
jl_split(x, "sentences")
returns three new rows with other variables duplicated, 
new 'tokens' values, and doc_ids 23.1 23.2 and 23.3
}
