% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_summarize_counts}
\alias{jl_summarize_counts}
\title{Collapse counts for grouped tibbles}
\usage{
jl_summarize_counts(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
an aggregated tibble
}
\description{
For grouped tibbles, collapse the 'counts' within each group. Like 
summarize, after group_by, but with no choice of function.
}
