% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_tokenize_categories}
\alias{jl_tokenize_categories}
\title{Create tokens using a dictionary}
\usage{
jl_tokenize_categories(x, dictionary, ...)
}
\arguments{
\item{x}{a tibble}

\item{dictionary}{a quanteda content analysis dictionary}

\item{...}{extra arguments to tokenizers::tokenizer_*}
}
\value{
a tibble
}
\description{
Runs a dictionary on the text column of x to create 'tokens' variable 
consisting of dictionary matches to each word. Non-matching words
from the dictionary generate no token.
}
