% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_tokenize_words}
\alias{jl_tokenize_words}
\title{Split text into words}
\usage{
jl_tokenize_words(x, ...)
}
\arguments{
\item{x}{a tibble}

\item{...}{extra arguments to tokenizers::tokenize_*}
}
\value{
a tibble
}
\description{
Fills the 'tokens' variable with word tokens.
}
