% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_types}
\alias{jl_types}
\title{Get the vocabulary under counts}
\usage{
jl_types(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a vector of type values
}
\description{
Returns the list of types that are counted in 'counts'. If 
'counts' does not exist, does the calculation for 'tokens'
}
