% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{K2009}
\alias{K2009}
\title{Interest Groups}
\description{
Interest Groups and the European Commission
}
\details{
Word counts from interest groups and a European Commission proposal to
reduce CO2 emissions in 2007.

\code{K2009} is a \code{jl_df} object.
}
\references{
H. Kluever (2009) 'Measuring influence group influence using
quantitative text analysis' European Union Politics 10(4) 535-549.
}
\keyword{datasets}
