% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{LB2002}
\alias{LB2002}
\title{The 1991 Irish Confidence debate}
\description{
Irish Confidence Debate (jl format)
}
\details{
This are word counts from the no-confidence motion debated in the
Irish Dáil from 16-18 October 1991 over the future of the Fianna 
Fail-Progressive Democrat coalition. 

\code{LB2003} is \code{jl_df} object.
}
\references{
Laver, M. & Benoit, K.R. (2002). Locating TDs in Policy Spaces:
            Wordscoring Dáil Speeches. Irish Political Studies, 17(1), 
            59–73.
}
\keyword{datasets}
