% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{LB2013}
\alias{LB2013}
\title{Irish Budget Debate Data 2009}
\description{
Irish budget debate 2009
}
\details{
These are word counts from the 2009 Budget debate in Ireland.

\code{LB2013} is a \code{jl_df} object
}
\references{
W. Lowe and K. Benoit (2013) 'Validating estimates of latent 
            traits from textual data using human judgment as a benchmark'
            Political Analysis 21(3) 298-313.
}
\keyword{datasets}
