% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{LG2000}
\alias{LG2000}
\title{UK Manifesto Data}
\description{
UK manifesto data from Laver et al.
}
\details{
This are word counts from the manifestos of the three main UK parties for
the 1992 and 1997 elections.

\code{LG2000} is a \code{jl_df} object.
}
\references{
M. Laver, K. Benoit and J. Garry (2003) 'Estimating policy 
            positions from political text using words as data' 
            American Political Science Review 97(2) 311-331.
}
\keyword{datasets}
