% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.wordscores.R
\name{plot.classic.wordscores}
\alias{plot.classic.wordscores}
\title{Plot a Wordscores Model}
\usage{
\method{plot}{classic.wordscores}(x, ...)
}
\arguments{
\item{x}{a fitted Wordscores model}

\item{...}{other arguments, passed to the dotchart command}
}
\value{
A plot of the wordscores in increasing order.
}
\description{
Plots Wordscores from a fitted Wordscores model
}
\author{
Will Lowe
}
\seealso{
\code{\link{classic.wordscores}}
}

