% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordfish.R
\name{plot.coef.wordfish}
\alias{plot.coef.wordfish}
\title{Plot the Word Parameters From a Wordfish Model}
\usage{
\method{plot}{coef.wordfish}(x, pch = 20, psi = TRUE, ...)
}
\arguments{
\item{x}{a fitted Wordfish model}

\item{pch}{Default is to use small dots to plot positions}

\item{psi}{whether to plot word fixed effects}

\item{...}{Any extra graphics parameters to pass in}
}
\value{
A plot of sorted beta and optionally psi parameters.
}
\description{
Plots sorted beta and optionally also psi parameters from a Wordfish model
}
\author{
Will Lowe
}
\seealso{
\code{\link{wordfish}}
}

