% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordfish.R
\name{plot.wordfish}
\alias{plot.wordfish}
\title{Plot a Wordfish Model}
\usage{
\method{plot}{wordfish}(x, truevals = NULL, level = 0.95, pch = 20, ...)
}
\arguments{
\item{x}{a fitted Wordfish model}

\item{truevals}{True document positions if known}

\item{level}{Intended coverage of confidence intervals}

\item{pch}{Default is to use small dots to plot positions}

\item{...}{Any extra graphics parameters to pass in}
}
\value{
A plot of sorted estimated document positions, with confidence
intervals and true document positions, if these are available.
}
\description{
Plots a fitted Wordfish model with confidence intervals
}
\author{
Will Lowe
}
\seealso{
\code{\link{wordfish}}
}

