% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.prop.crct.R
\name{power_prop_crct}
\alias{power_prop_crct}
\title{Power Calculations for Clustered Two-Sample Test for Proportions}
\usage{
power_prop_crct(n = NULL, p1 = NULL, p2 = NULL, sig.level = 0.05,
  power = NULL, rho = NULL, nclusters = NULL,
  alternative = c("two.sided", "one.sided"), strict = FALSE,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{n}{Number of observations}

\item{sig.level}{Significance level (Type I error probability)}

\item{power}{Power of test (1 minus Type II error probability)}

\item{alternative}{One- or two-sided test}

\item{p0}{Probability under the null}

\item{pa}{Probability under the alternative}
}
\value{
Object of class \code{power.htest}, a list of the arguments
(including the computed one) augmented with method and note elements.
}
\description{
Compute power of test, or determine parameters to obtain target power.
}
\details{
The procedure uses uniroot to find the root of a discontinuous function so
some errors may pop up due to the given setup that causes the root-finding
procedure to fail. Also, since exact binomial tests are used we have
discontinuities in the function that we use to find the root of but despite
this the function is usually quite stable.
}
\examples{



}
\seealso{
\code{\link{binom.test}}
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{htest}

