% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.wordscores.R
\name{predict.classic.wordscores}
\alias{predict.classic.wordscores}
\title{Predict New Document Positions}
\usage{
\method{predict}{classic.wordscores}(object, newdata = NULL,
  rescale = c("lbg", "none"), z = 0.95, ...)
}
\arguments{
\item{object}{Fitted wordscores model}

\item{newdata}{An object of class wfm in which to look for word counts to
predict document ideal points. If omitted, the reference documents are used.}

\item{rescale}{Rescale method for estimated positions.}

\item{z}{Notional confidence interval coverage}

\item{...}{further arguments (quietly ignored)}
}
\value{
\code{predict.wordscores} produces a vector of predicted document
positions and standard errors and confidence intervals.
}
\description{
Predicts positions of new documents from a fitted Wordscores model
}
\details{
This is the method described in Laver et al. 2003, including rescaling for
more than one virgin text.  Confidence intervals are not provided if
\code{rescale} is 'none'.
}
\author{
Will Lowe
}
\seealso{
\code{\link{classic.wordscores}}
}

