\name{predict.wordfish}
\alias{predict.wordfish}
\title{Predict Method for Wordfish}
\description{Predicts positions of new documents using a fitted Wordfish model}

\usage{
\method{predict}{wordfish}(object, newdata=NULL, se.fit=FALSE,
                 interval=c("none", "confidence"),
                 level=0.95, ...)
}
\arguments{
      \item{object}{A fitted wordfish model}
      \item{newdata}{An optional data frame or object of class wfm
	in which to look for word counts to predict document ideal
	points which to predict.  If omitted, the fitted values are used.}
      \item{se.fit}{A switch indicating if standard errors are required.}
      \item{interval}{Type of interval calculation}
      \item{level}{Tolerance/confidence level}
      \item{...}{further arguments passed to or from other methods.}
}
\value{
   \code{predict.wordfish} produces a vector of predictions or a matrix of
   predictions and bounds with column names `fit' and `se.fit', and
   with `lwr', and `upr' if `interval' is also set.
}
\details{
  Standard errors for document positions are generated by numerically
  inverting the relevant Hessians from the profile likelihood of the
  multinomial form of the model.
}
\references{
}
\seealso{\code{\link{wordfish}}}
\author{Will Lowe}
