% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_utils.R
\name{sample_covariates}
\alias{sample_covariates}
\title{Sample covariate data}
\usage{
sample_covariates(sample_size, n_covariates, n_normal, cov_mat = NULL)
}
\arguments{
\item{sample_size}{integer specifying the sample size}

\item{n_covariates}{integer specifying the number of covariates}

\item{n_normal}{integer specifying the number of normal covariates (must be <= n_convariates)}

\item{cov_mat}{correlation matrix}
}
\value{
a tibble data frame of size \code{sample_size} x \code{n_covariates}
}
\description{
Sample covariate data
}
