% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{SP2008_soc}
\alias{SP2008_soc}
\title{Societal sections of German Party Manifestos}
\source{
These data courtesy of S.-O. Proksch.
}
\description{
A word frequency matrix from the societal sections of German political party
manifestos from 1990-2005.
}
\details{
\code{SP2008_soc} is a \code{jl_df} object
}
\references{
J. Slapin and S.-O. Proksch (2008) 'A scaling model for
estimating time-series party positions from texts' American Journal of
Political Science 52(3), 705-722.
}
\keyword{datasets}
