% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wordfish.R
\name{summary.wordfish}
\alias{summary.wordfish}
\title{Summarize a Wordfish Model}
\usage{
\method{summary}{wordfish}(object, level = 0.95, ...)
}
\arguments{
\item{object}{fitted wordfish model}

\item{level}{confidence interval coverage}

\item{...}{extra arguments, e.g. level}
}
\value{
A data.frame containing estimated document position with standard
errors and confidence intervals.
}
\description{
Summarises estimated document positions from a fitted Wordfish model
}
\details{
if `level' is passed to the function, e.g. 0.95 for 95 percent confidence,
this generates the appropriate width intervals.
}
\author{
Will Lowe
}
\seealso{
\code{\link{wordfish}}
}

