% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfm.R
\name{trim}
\alias{trim}
\title{Trim a Word Frequency Data}
\usage{
trim(wfm, min.count = 5, min.doc = 5, sample = NULL, verbose = TRUE)
}
\arguments{
\item{wfm}{an object of class wfm, or a data matrix}

\item{min.count}{the smallest permissible word count}

\item{min.doc}{the fewest permissible documents a word can appear in}

\item{sample}{how many words to randomly retain}

\item{verbose}{whether to say what we did}
}
\value{
If \code{sample} is a number then this many words will be retained
after \code{min.doc} and \code{min.doc} filters have been applied.
}
\description{
Ejects low frequency observations and subsamples
}
\author{
Will Lowe
}
\seealso{
\code{\link{wfm}}
}

