% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/austin-package.R
\docType{data}
\name{ukmanif}
\alias{ukmanif}
\title{UK Manifesto Data}
\description{
UK manifesto data from Laver et al.
}
\details{
This are word counts from the manifestos of the three main UK parties for
the 1992 and 1997 elections.

ukmanif is a word frequency object.
}
\references{
Laver, Benoit and Garry (2003) `Estimating policy positions from
political text using words as data' American Political Science Review 97(2).
}
\keyword{datasets}

