% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{words}
\alias{words}
\alias{words<-}
\title{Extract Words}
\usage{
words(wfm)

words(wfm) <- value
}
\arguments{
\item{wfm}{an object of type wfm}

\item{value}{replacement if assignment}
}
\value{
A list of words.
}
\description{
Extracts the words from a wfm object
}
\author{
Will Lowe
}
\seealso{
\code{\link{wfm}}, \code{\link{docs}}
}

