% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfm.R
\name{as.docword}
\alias{as.docword}
\title{Extract a Document by Word Matrix}
\usage{
as.docword(wfm)
}
\arguments{
\item{wfm}{an object of class wfm}
}
\value{
a document by word count matrix
}
\description{
Extract a word count matrix with documents as rows and words as columns
}
\details{
This is a helper function for wfm objects.  Use it instead of manipulating
wfm object themselves.
}
\author{
Will Lowe
}
\seealso{
\code{\link{as.worddoc}}, \code{\link{wfm}}
}

