\name{as.wfm}
\alias{as.wfm}
\title{Coerce to a Word Frequency Matrix}
\description{Constructs a wfm object from various other kinds of objects}
\usage{as.wfm(mat, word.margin=1)}
\arguments{
  \item{mat}{a matrix of counts}
  \item{word.margin}{which margin of mat represents the words}
}
\value{an object of class wfm}

\seealso{\code{\link{wfm}}}
\author{Will Lowe}
