% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Austin-package.R
\docType{package}
\name{Austin}
\alias{Austin}
\alias{Austin-package}
\alias{Austin-package}
\title{Collection of miscellaneous useful and semi-useful functions}
\description{
Collection of miscellaneous useful and semi-useful functions and add-on
functions that enhances a number of existing packages and provides In
particular in relation to statistical genetics
}
\references{
Ekstrøm, C. (2017). The R Primer. Chapman & Hall, 2nd edition.
}
\author{
Claus Thorn Ekstrøm \email{ekstrom@sund.ku.dk}\cr Maintainer: Claus Thorn Ekstrøm
\email{ekstrom@sund.ku.dk}
}
\keyword{#}
\keyword{'}
\keyword{.registration}
\keyword{=}
\keyword{@}
\keyword{Austin,}
\keyword{From}
\keyword{L}
\keyword{Rcpp}
\keyword{TRUE}
\keyword{eDy}
\keyword{i}
\keyword{ib}
\keyword{mp}
\keyword{n}
\keyword{o}
\keyword{package}
\keyword{r}
\keyword{s}
\keyword{sourceCpp}
\keyword{t}
\keyword{u}
