\name{coef.classic.wordscores}
\alias{coef.classic.wordscores}
\title{Show Wordscores}
\description{Lists wordscores from a fitted Wordscores model.}
\usage{
\method{coef}{classic.wordscores}(object, ...)
}
\arguments{
  \item{object}{a fitted Wordscores model}
  \item{...}{extra arguments, currently unused}
}
\value{
  The wordscores 
}
\seealso{
  \code{\link{classic.wordscores}}
}
\author{Will Lowe}
