\name{daildata}
\alias{daildata}
\docType{data}
\description{Irish No-Confidence Debate}
\title{The Irish No-Confidence debate }
\usage{data(daildata)}
\details{
  This are word counts from the no-confidence debate in Ireland.

  daildata is a word frequency object.
}

\references{
  Benoit and Laver's Irish Political Studies piece. (fixme!)
}

\keyword{datasets}
