\name{demanif.foreign}
\alias{demanif.foreign}
\docType{data}
\title{Foreign Policy Sections of German Party Manifestos}
\description{
  A word frequency matrix from the foreign policy sections of German political party
  manifestos from 1990-2005. 
}
\usage{data(demanif.foreign)}
\details{
  demanif.foreign is word frequency matrix
}
\source{
   These data courtesy are of S.-O. Proksch.
}
\references{
  J. Slapin and S.-O. Proksch (2008) 'A scaling model for
  estimating time-series party positions from texts' American Journal
  of Political Science 52(3), 705-722.
}
\keyword{datasets}
