\name{demanif}
\alias{demanif}
\docType{data}
\title{German Party Manifesto Data}
\description{
  A random sample of words and their frequency in German political party
  manifestos from 1990-2005. 
}
\usage{data(demanif)}
\details{
  demanif is word frequency matrix
}
\source{
  Wordfish website (http://www.wordfish.org)
}
\references{
  J. Slapin and S.-O. Proksch (2008) 'A scaling model for
  estimating time-series party positions from texts' American Journal
  of Political Science 52(3), 705-722.
}
\keyword{datasets}
