% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfm.R
\name{docs}
\alias{docs}
\alias{docs<-}
\title{Extract Document Names}
\usage{
docs(wfm)

docs(wfm) <- value
}
\arguments{
\item{wfm}{an object of type wfm}

\item{value}{replacement if assignment}
}
\value{
A list of document names.
}
\description{
Extracts the document names from a wfm object.
}
\seealso{
\code{\link{wfm}}
}
\author{
Will Lowe
}
