\name{fitted.wordfish}
\alias{fitted.wordfish}
\title{Get Fitted Values from a Wordfish Model}
\description{
  Extracts the estimated word rates from a fitted Wordfish model
}
\usage{
\method{fitted}{wordfish}(object, ...)
}
\arguments{
  \item{object}{a fitted Wordfish model}
  \item{...}{Unused}
}
\value{
  Expected counts in the word frequency matrix
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{Will Lowe}
\keyword{}
