\name{iebudget2009}
\alias{iebudget2009}
\alias{iebudget2009cov}
\docType{data}
\description{Irish budget debate 2009}
\title{Irish Budget Debate Data 2009}
\usage{data(iebudget2009)}
\details{
  This are word counts from the 2009 Budget debate in Ireland.

  This is a word frequency nmatrix.  Loading this data also makes 
  available \code{iebudget2009cov} which contains covariates for the speakers.
}

\keyword{datasets}
