\name{interestgroups}
\alias{interestgroups}
\docType{data}
\description{Interest Groups and the European Commission }
\title{Interest Groups}
\usage{data(interestgroups)}
\details{
  Word counts from interest groups and a European commission proposal
  to reduce CO2 emmissions in 2007.

  \code{comm1} and \code{comm2} are the commission's proposal before and after
   the proposals of the interest groups.
}

\references{
  H. Kluever (2009) 'Measuring influence group influence using quantitative
  text analysis' European Union Politics 11:1.
}

\keyword{datasets}
