% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wfm.R
\name{is.wfm}
\alias{is.wfm}
\title{Checks for Word Frequency Matrix}
\usage{
is.wfm(x)
}
\arguments{
\item{x}{a matrix of counts}
}
\value{
Whether the object can be used as a Word Frequency Matrix
}
\description{
Checks whether an object is a Word Frequency Matrix
}
\author{
Will Lowe
}
\seealso{
\code{\link{wfm}}
}

