\name{is.wfm}
\alias{is.wfm}
\title{Checks for Word Frequency Matrix}
\description{Checks whether an object is a Word Frequency Matrix}
\usage{is.wfm(x)}
\arguments{
  \item{x}{a matrix of counts}
}
\value{Whether the object can be used as a Word Frequency Matrix}
\seealso{\code{\link{wfm}}}
\author{Will Lowe}
