% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_dfm}
\alias{jl_dfm}
\title{Extract a document feature matrix}
\usage{
jl_dfm(x, rownames_from = NULL, sparse = TRUE)
}
\arguments{
\item{x}{a tibble}

\item{rownames_from}{column of x to use as rownames}

\item{sparse}{whether to return a sparse (the default) or dense matrix}
}
\value{
Matrix::CsparseMatrix (or base::matrix if sparse = FALSE)
}
\description{
Constructs a document feature 
matrix using 'counts' with column names taken from the types attribute 
of 'counts' and row names from 'doc_id' or a user provided variable.
 
If sparse = FALSE (the default) then the result is a CsparseMatrix from the 
Matrix package and if TRUE, a dense base::matrix. Note that Non-sparse 
document feature matrices may become quite large.
}
