% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_filter}
\alias{jl_filter}
\title{Filter by document variable}
\usage{
jl_filter(x, ..., .preserve = FALSE)
}
\arguments{
\item{x}{a jl_df}

\item{...}{a logical expression specifying rows}

\item{.preserve}{whether to keep groups (default: FALSE)}
}
\value{
a jl_df
}
\description{
Filter documents using logical expressions, as you would in dplyr. This 
function will re-index 'counts' as necessary
}
