% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_identify}
\alias{jl_identify}
\title{Add a document identifier}
\usage{
jl_identify(x, v = NULL, drop_original = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{v}{variable to be used as a unique identifier (Default: one will be constructed)}

\item{drop_original}{whether to remove v after it is assigned to 'doc_id'}
}
\value{
a jl_df
}
\description{
Adds a variable 'doc_id' to uniquely identify each row. This identifier
may later, if jl_expand is used, this will be augmented with information 
about disaggregation level.
}
\details{
Note: v may be the name of a variable or an entirely new character variable
of the appropriate length and uniqueness.
}
