% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jl.R
\name{jl_promote_counts}
\alias{jl_promote_counts}
\title{Make counts real variables in wide form}
\usage{
jl_promote_counts(x, prefix = NULL)
}
\arguments{
\item{x}{a tibble with 'counts'}

\item{prefix}{what to prefix each counted element's value when it turns into a variable name}
}
\value{
a tibble with more columns, one for each counted type
}
\description{
Make counts real variables in wide form
}
