\name{plot.classic.wordscores}
\alias{plot.classic.wordscores}
\title{Plot a Wordscores Model}
\description{Plots Wordscores from a fitted Wordscores model}
\usage{
\method{plot}{classic.wordscores}(x, ...)
}

\arguments{
  \item{x}{a fitted Wordscores model}
  \item{...}{other arguments, passed to the dotchart command}
}
\value{
  A plot of the wordscores in increasing order.
}
\seealso{
  \code{\link{classic.wordscores}}
}
\author{Will Lowe}

