\name{summary.classic.wordscores}
\alias{summary.classic.wordscores}
\title{Summarize an Classic Wordscores Model}
\description{Summarises a Wordscores model}
\usage{
\method{summary}{classic.wordscores}(object, ...)
}
\arguments{
  \item{object}{a fitted wordscores model}
  \item{...}{extra arguments (currently ignored)}
}
\value{
  A summary of information about the reference documents used 
  to fit the model. 
}
\details{To see the wordscores, use \code{coef}.}

\author{Will Lowe}

