\name{ukmanif}
\alias{ukmanif}
\docType{data}
\description{UK manifesto data from Laver et al.}
\title{UK Manifesto Data}
\usage{data(ukmanif)}
\details{
  This are word counts from the manifestos of the three main UK parties
  for the 1992 and 1997 elections.

  ukmanif is a word frequency object.
}

\references{
  Laver, Benoit and Garry (2003) `Estimating policy positions from
  political text using words as data' American Political Science Review
  97(2).
}

\keyword{datasets}
