% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wfm.R
\name{wordmargin}
\alias{wordmargin}
\title{Which margin holds the words}
\usage{
wordmargin(x)
}
\arguments{
\item{x}{a word frequency matrix}
}
\value{
1 if words are rows and 2 if words are columns.
}
\description{
Checks which margin (rows or columns) of a Word Frequency Matrix holds the
words
}
\details{
Changing the wordmargin by assignment just swaps the dimnames
}
\author{
Will Lowe
}
\seealso{
\code{\link{wfm}}
}

